+++ daemons/controld/controld_fsa.h
@@ -16,7 +16,7 @@
 #  include <crm/common/mainloop.h>
 #  include <crm/cluster.h>
 #  include <crm/cluster/election.h>
-#  include <crm/common/ipcs.h>
+#  include <crm/common/ipcs_internal.h>
 
 /*! States the controller can be in */
 enum crmd_fsa_state {
