+++ daemons/controld/controld_messages.h
@@ -1,5 +1,7 @@
 /*
- * Copyright 2004-2018 Andrew Beekhof <andrew@beekhof.net>
+ * Copyright 2004-2020 the Pacemaker project contributors
+ *
+ * The version control history for this file may have further details.
  *
  * This source code is licensed under the GNU Lesser General Public License
  * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
@@ -9,7 +11,7 @@
 #  define XML_CRM_MESSAGES__H
 
 #  include <crm/crm.h>
-#  include <crm/common/ipcs.h>
+#  include <crm/common/ipcs_internal.h>
 #  include <crm/common/xml.h>
 #  include <crm/cluster/internal.h>
 #  include <controld_fsa.h>
