+++ daemons/fenced/fenced_commands.c
@@ -25,7 +25,7 @@
 #include <crm/crm.h>
 #include <crm/msg_xml.h>
 #include <crm/common/ipc.h>
-#include <crm/common/ipcs.h>
+#include <crm/common/ipcs_internal.h>
 #include <crm/cluster/internal.h>
 #include <crm/common/mainloop.h>
