+++ daemons/schedulerd/pacemaker-schedulerd.c
@@ -21,7 +21,7 @@
 
 #include <libxml/parser.h>
 
-#include <crm/common/ipcs.h>
+#include <crm/common/ipcs_internal.h>
 #include <crm/common/mainloop.h>
 #include <crm/pengine/internal.h>
 #include <pacemaker-internal.h>
