+++ lib/lrmd/lrmd_client.c
@@ -26,7 +26,7 @@
 #include <crm/lrmd.h>
 #include <crm/services.h>
 #include <crm/common/mainloop.h>
-#include <crm/common/ipcs.h>
+#include <crm/common/ipcs_internal.h>
 #include <crm/common/remote_internal.h>
 #include <crm/msg_xml.h>
