+++ maint/mocked/based.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the Pacemaker project contributors
+ * Copyright 2019-2020 the Pacemaker project contributors
  *
  * The version control history for this file may have further details.
  *
@@ -11,7 +11,7 @@
 #include <stdlib.h>  /* size_t */
 #include <stdbool.h>  /* bool */
 
-#include <crm/common/ipcs.h>  /* crm_client_t */
+#include <crm/common/ipcs_internal.h>  /* crm_client_t */
 
 
 struct module_s;
